+++ elements/pfe-label/BaseLabel.ts
@@ -1,20 +1,29 @@
-import { html } from 'lit';
-import { customElement, property } from 'lit/decorators.js';
+import { LitElement, html } from 'lit';
+import { property } from 'lit/decorators.js';
+import { classMap } from 'lit/directives/class-map.js';
 
-import '@patternfly/pfe-button';
+import { SlotController } from '@patternfly/pfe-core/controllers/slot-controller.js';
 
-import { BaseLabel } from './BaseLabel.js';
+import styles from './BaseLabel.scss';
 
-import styles from './pfe-label.scss';
-import { ComposedEvent } from '@patternfly/pfe-core';
+export type LabelVariant = (
+  | 'filled'
+  | 'outline'
+);
 
+export type LabelColor = (
+  | 'blue'
+  | 'cyan'
+  | 'green'
+  | 'orange'
+  | 'purple'
+  | 'red'
+  | 'grey'
+  | 'gold'
+)
 
 /**
- * Labels allow users to display meta data in a stylized form.
- *
- * @summary Allows users to display meta data in a stylized form.
- *
- * @fires close - when a removable label's close button is clicked
+ * Base label class
  *
  * @cssprop {<length>} --pf-c-label--FontSize   {@default `0.875em`}
  *
@@ -86,56 +95,59 @@ import { ComposedEvent } from '@patternfly/pfe-core';
  *
  * @slot
  *       Must contain the text for the label.
- *
- * @cssprop {<length>} --pf-c-label--m-compact--PaddingTop     {@default `0`}
- * @cssprop {<length>} --pf-c-label--m-compact--PaddingRight   {@default `0.5rem`}
- * @cssprop {<length>} --pf-c-label--m-compact--PaddingBottom  {@default `0`}
- * @cssprop {<length>} --pf-c-label--m-compact--PaddingLeft    {@default `0.5rem`}
- */
-@customElement('pfe-label')
-export class PfeLabel extends BaseLabel {
-  static readonly version = '{{version}}';
+*/
 
-  static readonly styles = [...BaseLabel.styles, styles];
+export abstract class BaseLabel extends LitElement {
+  static readonly styles = [styles];
 
-  static readonly shadowRootOptions: ShadowRootInit = { ...BaseLabel.shadowRootOptions, delegatesFocus: true };
+  /**
+   * Changes the style of the label.
+   * - Filled: Colored background with colored border.
+   * - Outline: White background with colored border.
+   */
+  @property({ reflect: true }) variant: LabelVariant = 'filled';
 
-  /** Flag indicating the label is compact */
-  @property({ reflect: true, type: Boolean }) compact = false;
+  /**
+   * Changes the color of the label
+   */
+  @property({ reflect: true }) color: LabelColor = 'grey';
 
-  /** Flag indicating the label text should be truncated */
-  @property({ reflect: true, type: Boolean }) truncated = false;
+  /** Shorthand for the `icon` slot, the value is icon name */
+  @property() icon = '';
 
-  /** Flag indicating the label is removable */
-  @property({ reflect: true, type: Boolean }) removable = false;
+  /** Represents the state of the anonymous and icon slots */
+  protected slots = new SlotController(this, null, 'icon');
 
-  /** Text label for a removable label's close button */
-  @property({ attribute: 'close-button-label' }) closeButtonLabel?: string;
-
-  protected override renderDefaultIcon() {
-    return !this.icon ? '' : html`
-      <pfe-icon ?hidden=${!this.icon} icon=${this.icon} size="sm"></pfe-icon>
+  override render() {
+    const { icon } = this;
+    const hasIcon = this.slots.hasSlotted('icon') || !!icon;
+    return html`
+      <span id="container" class=${classMap({ hasIcon })}>
+        <span part="icon">
+          <slot name="icon">${this.renderDefaultIcon()}</slot>
+        </span>
+        <span id="text">
+          <slot></slot>
+        </span>
+      </span>${this.renderSuffix?.() ?? ''}
     `;
   }
 
-  protected override renderSuffix() {
-    return !this.removable ? '' : html`
-      <span part="close-button" ?hidden=${!this.removable}>
-        <pfe-button plain @click=${() => this.dispatchEvent(new ComposedEvent('close'))}>
-          <button>
-            <svg slot="icon" style="vertical-align:-0.125em" fill="currentColor" height="1em" width="1em" viewBox="0 0 352 512">
-              <title>${this.closeButtonLabel ?? 'remove'}</title>
-              <path d="M242.72 256l100.07-100.07c12.28-12.28 12.28-32.19 0-44.48l-22.24-22.24c-12.28-12.28-32.19-12.28-44.48 0L176 189.28 75.93 89.21c-12.28-12.28-32.19-12.28-44.48 0L9.21 111.45c-12.28 12.28-12.28 32.19 0 44.48L109.28 256 9.21 356.07c-12.28 12.28-12.28 32.19 0 44.48l22.24 22.24c12.28 12.28 32.2 12.28 44.48 0L176 322.72l100.07 100.07c12.28 12.28 32.2 12.28 44.48 0l22.24-22.24c12.28-12.28 12.28-32.19 0-44.48L242.72 256z"/>
-            </svg>
-          </button>
-        </pfe-button>
-      </span>
-    `;
-  }
-}
+  /**
+   * Fallback content for the icon slot. When the `icon` attribute is set, it
+   * should render an icon corresponding to the value.
+   *
+   * @example ```html
+   * <pfe-icon icon=${this.icon}></pfe-icon>
+   * ```
+   */
+  protected abstract renderDefaultIcon(): unknown;
 
-declare global {
-  interface HTMLElementTagNameMap {
-    'pfe-label': PfeLabel;
-  }
+  /**
+   * Optional override to render content after the anonymous slot.
+   * @example ```html
+   * <button id="close-button">X</button>
+   * ```
+   */
+  protected abstract renderSuffix?(): unknown;
 }