+++ elements/pfe-button/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @patternfly/pfe-button
 
+## 2.0.0-next.4
+
+### Major Changes
+
+- a730678c: Several changes align `<pfe-button>` to [PatternFly v4](https://patternfly.org/components/button).
+
+  NEW:
+  `warning`, `link`, and `control` variants added.
+  `icon` and `loading` attributes added.
+  `plain`, `block`, `warning` and `loading` attributes added.
+  `icon` slot added.
+
+  BREAKING CHANGES:
+  The `pfelement` attribute and `PFElement` class are **removed** from the `<pfe-button>` element by default.
+  The `danger` variant is **removed** in favour of a new `danger` variant.
+  _All_ the `--pfe-*` css variables are **removed** in favour of their `--pf-*` equivalents.
+
+  See the [docs](https://patternflyelements.org/components/button) for more info,
+  and the [demo](https://patternflyelements.org/components/button/demo) for usage examples.
+
+### Patch Changes
+
+- Updated dependencies [34ecd410]
+  - @patternfly/pfe-core@2.0.0-next.6
+
 ## 2.0.0-next.3
 
 ### Patch Changes